requirements.txt
# Google Cloud & Vertex AI SDKs
google-cloud-aiplatform
google-cloud-firestore
google-generativeai
google-auth
google-auth-oauthlib

# Core Libraries
pandas
numpy
